requirements.txt
pandas
pyyaml
scikit-learn
mlflow
urllib3